Sensible renaming and extracting functions to different files.
Execution reconciler still not working as expected in cluster but it is
working in local environment.

It may not be an absolute fix but it's a step in the right direction.
